ingest-router: add executor and actually return results to client (#93)
What this does:

- introduces the executor. this module splits the request, calls
upstreams, enforces timeouts, and merges the responses back for the
caller

- wires up the executor to the the ingest router service so the service
now returns real results from upstreams.

- adds a test that spins up the mock relay api to demonstrate the
project config endpoint working end to end

This PR is built on top of https://github.com/getsentry/synapse/pull/92
-- please review that one first

+++ ingest-router/src/config.rs
@@ -66,17 +66,17 @@ pub struct RelayTimeouts {
     /// HTTP request timeout for individual upstream calls (seconds).
     /// This is the maximum time a single HTTP request can take.
     /// Default: 15 seconds
-    pub http_timeout_secs: u16,
+    pub http_timeout_secs: u64,
 
     /// Task timeout when waiting for the first upstream to respond (seconds).
     /// Must be >= http_timeout_secs to allow HTTP requests to complete.
     /// Default: 20 seconds
-    pub task_initial_timeout_secs: u16,
+    pub task_initial_timeout_secs: u64,
 
     /// Deadline for all remaining tasks after first success (seconds).
     /// Aggressively cuts off slow upstreams once we have good data.
     /// Default: 5 seconds
-    pub task_subsequent_timeout_secs: u16,
+    pub task_subsequent_timeout_secs: u64,
 }
 
 impl Default for RelayTimeouts {
